Changes in files and ErrorMessage
ErrorMessage => Message
Some directory organisational changes
