+++ src/CSS/main.css
@@ -5,17 +5,3 @@ body,
   width: 100%;
   height: 100%;
 }
-
-.error-message {
-  background-color: #fb9898;
-  border-style: solid;
-  border-color: #fc3f3f;
-  border-width: thin;
-  border-radius: 12px;
-  text-align: center;
-  font-size: x-large;
-  width: 100%;
-  opacity: 1;
-  padding: 10px;
-  outline: none;
-}
